refactored data folder structure; added data view to dummy data maker
