+++ package.json
@@ -23,6 +23,7 @@
     "pg-hstore": "^2.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-infinite-scroll-component": "^6.1.0",
     "react-microsoft-login": "^2.0.1",
     "react-pro-sidebar": "^1.1.0",
     "react-router-dom": "^6.22.1",
