+++ src/components/scheduleTable.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "../styles/components/scheduleTable.module.scss";
 import {useState, useEffect} from "react";
-import Data from "../dummydata/dumbdata.csv";
+import Data from "../data/dummydata/dumbdata.csv";
 import Papa from "papaparse"
 
 export default function Scheduletable()
