+++ src/data/dummydata/styles/components/dataview.module.scss
@@ -1,0 +1,4 @@
+.DataView {
+  display: flex;
+  flex-direction: column;
+}
